@@ -41,7 +41,7 @@ class Plugin implements BundlePluginInterface, RoutingPluginInterface
     /**
      * {@inheritdoc}
      */
-    public function getBundles(ParserInterface $parser)
+    public function getBundles(ParserInterface $parser): array
     {
         return [
             BundleConfig::create(SyncCtoProBundle::class)
@@ -67,6 +67,7 @@ class Plugin implements BundlePluginInterface, RoutingPluginInterface
     {
         return $resolver
             ->resolve(__DIR__ . '/../Resources/config/routing.yml')
-            ->load(__DIR__ . '/../Resources/config/routing.yml');
+            ->load(__DIR__ . '/../Resources/config/routing.yml')
+        ;
     }
 }
